Do not change position when changing chapter
When paging to the previous chapter with "b", it's useful to jump to the
bottom (the bottom is where the previous "page" is). However, when
jumping to the previous chapter with "H", I think it makes more sense to
jump to the top (so things like the chapter title and other clues are
visible).