fix: make page stats more consistent
Previously, the page index would sometimes make unexpected jumps. To fix
this, better line offset management was needed.

Internally, tview.TextView keeps its vertical offset within the range
0-(linecount-height) inside its Draw() function. Since we use that
offset before Draw() gets called (e.g. to calculate page stats), we also
need to make sure it stays within that range.

Additionally, due to this offset range limitation, the last line of the
final page stays at the bottom of the viewport when scrolled to the end.
This makes page counting ambiguous if the last page is short. For
example, we can reach the end of the chapter with 70 lines of page 4
visible and 10 lines of page 5 visible. Do we display a page index of 4
or 5 in this scenario? Ideally, I think the page comprising most of the
viewport would determine the page index to show, but this isn't possible
given the offset range constraints.

To disambiguate, the page index is tracked using the top of a page. In
the above scenario, for example, as soon as the first line of page 5
enters the viewport, we will display 5 as the page index. The trade-off
is that scrolling down even a single line from the first page will
switch the page index from 1 to 2.